+++ src/config/interceptor.js
@@ -1,0 +1,13 @@
+import axios from 'axios';
+
+axios.interceptors.request.use((config) => {
+    const newConfig = {
+      ...config,
+      headers: {
+        ...config.headers,
+        env: process.env.REACT_APP_ENV || 'development',
+      },
+    };
+  
+    return newConfig;
+  });
